# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TradeHub - Deriv Companion API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List connected Deriv accounts
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: connectAccount
      tags: [accounts]
      summary: Connect a Deriv account using an API token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectAccountBody"
      responses:
        "201":
          description: Account connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
...
  /accounts/{id}:
    delete:
      operationId: disconnectAccount
      tags: [accounts]
      summary: Disconnect (remove) an account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Disconnected
    patch:
      operationId: updateAccount
      tags: [accounts]
      summary: Update an account (set as active, rename label)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountBody"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{id}/refresh:
    post:
      operationId: refreshAccountBalance
      tags: [accounts]
      summary: Pull the latest realtime balance for an account from Deriv
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account with refreshed balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /bots:
    get:
      operationId: listBots
      tags: [bots]
      summary: List imported bots
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: Search by name or tag
      responses:
        "200":
          description: List of bots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bot"
    post:
      operationId: importBot
      tags: [bots]
      summary: Import a new bot (XML content)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportBotBody"
      responses:
        "201":
          description: Bot imported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
  /bots/{id}:
    get:
      operationId: getBot
      tags: [bots]
      summary: Get bot details (including XML content)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotDetail"
    patch:
      operationId: updateBot
      tags: [bots]
      summary: Update bot metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBotBody"
      responses:
        "200":
          description: Updated bot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
    delete:
      operationId: deleteBot
      tags: [bots]
      summary: Delete a bot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /bots/{id}/toggle-favorite:
    post:
      operationId: toggleBotFavorite
      tags: [bots]
      summary: Toggle the favorite flag for a bot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated bot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
...
  /tools/martingale:
    post:
      operationId: calculateMartingale
      tags: [tools]
      summary: Calculate a martingale staking plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MartingaleBody"
      responses:
        "200":
          description: Martingale plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MartingaleResult"
  /tools/risk:
    post:
      operationId: calculateRisk
      tags: [tools]
      summary: Position size based on account risk percentage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiskBody"
      responses:
        "200":
          description: Risk plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskResult"
  /tools/compound:
    post:
      operationId: calculateCompound
      tags: [tools]
      summary: Compound growth projection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompoundBody"
      responses:
        "200":
          description: Compound projection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompoundResult"
...
components:
  schemas:
...
    Account:
      type: object
      properties:
        id:
          type: integer
        label:
          type: string
        loginid:
          type: string
        accountType:
          type: string
          description: real or demo
        currency:
          type: string
        balance:
          type: number
        email:
          type: string
        country:
          type: string
        isActive:
          type: boolean
        connectedAt:
          type: string
          format: date-time
      required: [id, label, loginid, accountType, currency, balance, isActive, connectedAt]
    ConnectAccountBody:
      type: object
      properties:
        label:
          type: string
          minLength: 1
        apiToken:
          type: string
          minLength: 8
          description: Deriv API token (read/trade scopes)
      required: [label, apiToken]
    UpdateAccountBody:
      type: object
      properties:
        label:
          type: string
        isActive:
          type: boolean
...
    Bot:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        strategy:
          type: string
          description: e.g. Martingale, D'Alembert, Custom
        market:
          type: string
          description: e.g. Volatility 75, Boom 1000
        tags:
          type: array
          items:
            type: string
        favorite:
          type: boolean
        sizeBytes:
          type: integer
        status:
          type: string
          enum: [idle, running, paused]
        lastRunAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, name, strategy, market, tags, favorite, sizeBytes, status, createdAt]
    BotDetail:
      allOf:
        - $ref: "#/components/schemas/Bot"
        - type: object
          properties:
            xmlContent:
              type: string
              description: Full Deriv DBot XML content
          required: [xmlContent]
    ImportBotBody:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        strategy:
          type: string
        market:
          type: string
        tags:
          type: array
          items:
            type: string
        xmlContent:
          type: string
          minLength: 1
      required: [name, xmlContent]
    UpdateBotBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        strategy:
          type: string
        market:
          type: string
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [idle, running, paused]
...
    MartingaleBody:
      type: object
      properties:
        baseStake:
          type: number
          minimum: 0.01
        multiplier:
          type: number
          minimum: 1
        steps:
          type: integer
          minimum: 1
          maximum: 25
        payoutPercent:
          type: number
          description: e.g. 95 for 95% payout
      required: [baseStake, multiplier, steps, payoutPercent]
    MartingaleStep:
      type: object
      properties:
        step:
          type: integer
        stake:
          type: number
        cumulativeLoss:
          type: number
        payout:
          type: number
        netIfWin:
          type: number
      required: [step, stake, cumulativeLoss, payout, netIfWin]
    MartingaleResult:
      type: object
      properties:
        steps:
          type: array
          items:
            $ref: "#/components/schemas/MartingaleStep"
        totalRisk:
          type: number
      required: [steps, totalRisk]
    RiskBody:
      type: object
      properties:
        accountBalance:
          type: number
          minimum: 0
        riskPercent:
          type: number
          minimum: 0.1
          maximum: 100
        stopLossPips:
          type: number
          minimum: 1
        pipValue:
          type: number
          minimum: 0.01
      required: [accountBalance, riskPercent, stopLossPips, pipValue]
    RiskResult:
      type: object
      properties:
        riskAmount:
          type: number
        positionSize:
          type: number
      required: [riskAmount, positionSize]
    CompoundBody:
      type: object
      properties:
        startingBalance:
          type: number
          minimum: 0
        dailyReturnPercent:
          type: number
        days:
          type: integer
          minimum: 1
          maximum: 365
      required: [startingBalance, dailyReturnPercent, days]
    CompoundPoint:
      type: object
      properties:
        day:
          type: integer
        balance:
          type: number
      required: [day, balance]
    CompoundResult:
      type: object
      properties:
        points:
          type: array
          items:
            $ref: "#/components/schemas/CompoundPoint"
        finalBalance:
          type: number
        totalProfit:
          type: number
      required: [points, finalBalance, totalProfit]
...