# lib/api-zod/src/generated/types/martingaleBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TradeHub - Deriv Companion API
 * OpenAPI spec version: 0.1.0
 */

export interface MartingaleBody {
  /** @minimum 0.01 */
  baseStake: number;
  /** @minimum 1 */
  multiplier: number;
  /**
   * @minimum 1
   * @maximum 25
   */
  steps: number;
  /** e.g. 95 for 95% payout */
  payoutPercent: number;
}