# lib/api-zod/src/generated/types/riskBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TradeHub - Deriv Companion API
 * OpenAPI spec version: 0.1.0
 */

export interface RiskBody {
  /** @minimum 0 */
  accountBalance: number;
  /**
   * @minimum 0.1
   * @maximum 100
   */
  riskPercent: number;
  /** @minimum 1 */
  stopLossPips: number;
  /** @minimum 0.01 */
  pipValue: number;
}