# lib/api-zod/src/generated/types/riskResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TradeHub - Deriv Companion API
 * OpenAPI spec version: 0.1.0
 */

export interface RiskResult {
  riskAmount: number;
  positionSize: number;
}