# lib/api-zod/src/generated/types/updateAccountBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TradeHub - Deriv Companion API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateAccountBody {
  label?: string;
  isActive?: boolean;
}