# lib/api-zod/src/src/generated/types/account.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TradeHub - Deriv Companion API
 * OpenAPI spec version: 0.1.0
 */

export interface Account {
  id: number;
  label: string;
  loginid: string;
  /** real or demo */
  accountType: string;
  currency: string;
  balance: number;
  email?: string;
  country?: string;
  isActive: boolean;
  connectedAt: Date;
}