# lib/api-zod/src/src/generated/types/bot.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TradeHub - Deriv Companion API
 * OpenAPI spec version: 0.1.0
 */
import type { BotStatus } from "./botStatus";

export interface Bot {
  id: number;
  name: string;
  description?: string;
  /** e.g. Martingale, D'Alembert, Custom */
  strategy: string;
  /** e.g. Volatility 75, Boom 1000 */
  market: string;
  tags: string[];
  favorite: boolean;
  sizeBytes: number;
  status: BotStatus;
  lastRunAt?: Date | null;
  createdAt: Date;
}