# lib/api-zod/src/src/generated/types/martingaleResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TradeHub - Deriv Companion API
 * OpenAPI spec version: 0.1.0
 */
import type { MartingaleStep } from "./martingaleStep";

export interface MartingaleResult {
  steps: MartingaleStep[];
  totalRisk: number;
}